+++ Flow/src/utopia/flow/event/listener/ConditionalChangeReaction.scala
@@ -1,9 +1,9 @@
 package utopia.flow.event.listener
 
-import utopia.flow.event.model.ChangeResponse.{Continue, Detach}
+import utopia.flow.event.model.ChangeResponse.Detach
 import utopia.flow.event.model.{ChangeEvent, ChangeResponse}
-import utopia.flow.operator.enumeration.End.{First, Last}
 import utopia.flow.operator.enumeration.End
+import utopia.flow.operator.enumeration.End.{First, Last}
 import utopia.flow.view.mutable.Pointer
 import utopia.flow.view.mutable.eventful.ResettableFlag
 import utopia.flow.view.template.eventful.Changing
@@ -198,7 +198,7 @@ class ConditionalChangeReaction[A](origin: Changing[A], conditionPointer: Changi
 			else
 				detachmentQueuedFlag.set()
 			
-			ChangeResponse.continueIf(!ended && origin.isChanging)
+			ChangeResponse.continueIf(!ended && origin.mayChange)
 		}
 	}